requirements.txt
Django==2.2.19
django-tailwind==0.6.1
djangorestframework==3.12.2
nodeenv==1.5.0
npm==0.1.1
optional-django==0.1.0
pytz==2021.1
sqlparse==0.4.1
